refractor process view into cleaner UI UX
