+++ src/components/icons/index.ts
@@ -14,3 +14,4 @@ export * from './SparklesIcon';
 export * from './VolumeIcon';
 export * from './SquareStopIcon';
 export * from './ChevronDownIcon';
+export * from './TracesIcon';
